% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASSO_lmls.R
\name{lmls}
\alias{lmls}
\title{Gaussian location-scale regression}
\usage{
lmls(
  location,
  scale = ~1,
  data = environment(location),
  lasso = FALSE,
  lambda = NULL,
  const = NULL,
  light = TRUE,
  maxit = 100,
  reltol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{location}{A two-sided formula with the response variable on the LHS
and the predictor for the mean on the RHS.}

\item{scale}{A one-sided formula with the predictor for the standard
deviation on the RHS.}

\item{data}{A data frame (or list or environment) in which to evaluate
the \code{location} and \code{scale} formulas.}

\item{lasso}{A logical values that implies weather the model will be penalized with
a L1 regularization.}

\item{lambda}{A list of values that are used as the lambdas in the penalization
for location and scale coefficients.}

\item{const}{A small constant used in the quadratic approximation for the
penalization.}

\item{light}{If \code{TRUE}, the design matrices are removed from the estimated
model to save some memory.}

\item{maxit}{The maximum number of iterations of the Fisher scoring
algorithm.}

\item{reltol}{The relative convergence tolerance of the Fisher scoring
algorithm.}
}
\value{
A fitted linear model for location and scale as an \code{lmls} S3 object.
The object has at least the following entries:
\itemize{
\item \code{y}: the response vector
\item \code{nobs}: the number of observations
\item \code{df}: the degrees of freedom
\item \code{df.residual}: the residual degrees of freedom
\item \code{coefficients}: the regression coefficients as a list with the names
\code{location} and \code{scale}
\item \code{fitted.values}: the fitted values as a list with the names \code{location}
and \code{scale}
\item \code{residuals}: the response residuals
\item \code{coefficients}: the variance-covariance matrices of the regression
coefficients as a list with the names \code{location} and \code{scale}
\item \code{iterations}: the number of iterations the Fisher scoring algorithm
took to converge
}
}
\description{
The location-scale regression model assumes a normally distributed response
variable with one linear predictor for the mean (= the location) and one for
the standard deviation (= the scale). The standard deviation is mapped to
the linear predictor through a log link.

This function sets up the model object and estimates it with maximum
likelihood.
}
\examples{
library(asp22lasso)
n <- 1000
x1 <- runif(n)
x2 <- runif(n)
x3 <- runif(n)
y <- rnorm(n, 0 + 1 * x1 + 1 * x3, exp(-3 + 1 * x2 + 1 * x3))
data <- data.frame(matrix(c(y, x1, x2, x3), ncol = 4))
names(data) <- c("y", "X1", "X2", "X3")
m <- lmls(y ~ X1 + X3, ~ X2 + X3, data = data)
summary(m)
plot(m)
qqnorm(m)
}
