% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.lmls}
\alias{summary.lmls}
\title{Summary for LMLS}
\usage{
\method{summary}{lmls}(
  object,
  type = c("ml", "boot", "mcmc"),
  digits = max(3, getOption("digits") - 3),
  ...
)
}
\arguments{
\item{object}{A location-scale regression model from the \code{\link[=lmls]{lmls()}} function.}

\item{type}{\code{"ml"}:
\itemize{
\item If \code{"ml"}, the maximum likelihood estimates and the
asymptotic standard errors are shown.
}}

\item{digits}{The number of digits to print.}

\item{...}{Passed on to \code{\link[=printCoefmat]{printCoefmat()}}.}
}
\value{
The (unmodified and invisible) \code{lmls} S3 object, see \code{\link[=lmls]{lmls()}}.
}
\description{
Prints a summary for location-scale regression models from the \code{\link[=lmls]{lmls()}} and
\code{\link[=lasso_lmls]{lasso_lmls()}} function.
}
